requirements.txt
asgiref==3.4.1
black==21.12b0
click==8.0.3
Django==4.0.1
django-js-asset==1.2.2
django-mptt==0.13.4
djangorestframework==3.13.1
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.0
platformdirs==2.4.1
pytz==2021.3
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
